refactor: Move the tag mask constants to the Bot file
- More consistent since other bot-related constants are stored there.
